workflow-validator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.6.1
python-dotenv==1.0.0
